Directed edge arrow drawing improved
In the python interface, drawing/edge.py is modified. Now, the tip
of the arrow head is placed precisely where the Bezier curve intersects
the the (circular) vertex. This can't be solved exactly, so a numerical
Bisection solver is implemented for this purpose.

Furthermore, the curved edge is drawn up to the base of the arrowhead,
and not all the way to the center of the destination vertex. The
Bezier curve is adjusted such that the curved edge blends in with
the arrow head.