setup.py: changed build folder of igraph to vendor/build/igraph
This folder is used when igraph is downloaded and compiled during the
invocation of setup.py. The whole thing might be deprecated in the
future, but currently my focus is to switch to git submodules and the
first step is to clean up the root folder a bit.

As a side effect, the entire setup.py was reformatted with Black.