fix: handle PathLike in Graph.write and Graph.Read
os.PathLike was introduced in Python 3.6, and given igraph's aim of
supporting the three most recent minor versions of Python 3, depending
on it should not be an issue.

This is a minimal fix, which simply turns PathLike objects into their
string representation and otherwise keeps existing behavior. A more
ambitious fix would be to rewrite how files and paths are handled using
pathlib.Path.